pyproject.toml

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [
  {name = "ninanor", email = "ninanor@nina.no"}
]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = [
  "arro3-core>=0.8.0",
  "dash>=4.0.0",
  "dash-ag-grid>=33.3.3",
  "django-environ",
  "fsspec>=2026.2.0",
  "hydra-core>=1.3.2",
  "obstore>=0.9.2",
  "structlog>=25.5.0",
  "violit>=0.4.0",
]
description = ""
license = "GPL-3.0+"
name = "s3-explorer"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"


[project.scripts]
s3-explorer = "s3_explorer.main:start"

[dependency-groups]
dev = [
  "deptry",
  "hatchling",
]

[tool.ruff]
fix = true

[tool.ruff.lint]
ignore = ["COM812"]
select = ["E", "W", "I", "F", "UP", "S", "B", "A", "COM", "LOG", "PTH", "Q"]
